requirement graph.txt
auto-py-to-exe==2.41.0
├── Eel [required: >=0.11.0, installed: 0.16.0]
│   ├── bottle [required: Any, installed: 0.12.25]
│   ├── bottle-websocket [required: Any, installed: 0.2.9]
│   │   ├── bottle [required: Any, installed: 0.12.25]
│   │   └── gevent-websocket [required: Any, installed: 0.10.1]
│   │       └── gevent [required: Any, installed: 23.9.1]
│   │           ├── cffi [required: >=1.12.2, installed: 1.16.0]
│   │           │   └── pycparser [required: Any, installed: 2.21]
│   │           ├── greenlet [required: >=3.0rc3, installed: 3.0.0]
│   │           ├── zope.event [required: Any, installed: 5.0]
│   │           │   └── setuptools [required: Any, installed: 68.2.0]
│   │           └── zope.interface [required: Any, installed: 6.1]
│   │               └── setuptools [required: Any, installed: 68.2.0]
│   ├── future [required: Any, installed: 0.18.3]
│   ├── pyparsing [required: Any, installed: 3.1.1]
│   └── whichcraft [required: Any, installed: 0.6.1]
└── pyinstaller [required: >=5.8.0, installed: 6.0.0]
    ├── altgraph [required: Any, installed: 0.17.4]
    ├── packaging [required: >=20.0, installed: 23.2]
    ├── pefile [required: >=2022.5.30, installed: 2023.2.7]
    ├── pyinstaller-hooks-contrib [required: >=2021.4, installed: 2023.9]
    ├── pywin32-ctypes [required: >=0.2.1, installed: 0.2.2]
    └── setuptools [required: >=42.0.0, installed: 68.2.0]
PySide6==6.5.3
├── PySide6-Addons [required: ==6.5.3, installed: 6.5.3]
│   ├── PySide6-Essentials [required: ==6.5.3, installed: 6.5.3]
│   │   └── shiboken6 [required: ==6.5.3, installed: 6.5.3]
│   └── shiboken6 [required: ==6.5.3, installed: 6.5.3]
├── PySide6-Essentials [required: ==6.5.3, installed: 6.5.3]
│   └── shiboken6 [required: ==6.5.3, installed: 6.5.3]
└── shiboken6 [required: ==6.5.3, installed: 6.5.3]
torchaudio==2.1.0
└── torch [required: ==2.1.0, installed: 2.1.0]
    ├── filelock [required: Any, installed: 3.9.0]
    ├── fsspec [required: Any, installed: 2023.4.0]
    ├── jinja2 [required: Any, installed: 3.1.2]
    │   └── MarkupSafe [required: >=2.0, installed: 2.1.2]
    ├── networkx [required: Any, installed: 3.0]
    ├── sympy [required: Any, installed: 1.12]
    │   └── mpmath [required: >=0.19, installed: 1.3.0]
    └── typing-extensions [required: Any, installed: 4.4.0]
torchvision==0.16.0
├── numpy [required: Any, installed: 1.24.1]
├── pillow [required: >=5.3.0,!=8.3.*, installed: 9.3.0]
├── requests [required: Any, installed: 2.28.1]
│   ├── certifi [required: >=2017.4.17, installed: 2022.12.7]
│   ├── charset-normalizer [required: >=2,<3, installed: 2.1.1]
│   ├── idna [required: >=2.5,<4, installed: 3.4]
│   └── urllib3 [required: >=1.21.1,<1.27, installed: 1.26.13]
└── torch [required: ==2.1.0, installed: 2.1.0]
    ├── filelock [required: Any, installed: 3.9.0]
    ├── fsspec [required: Any, installed: 2023.4.0]
    ├── jinja2 [required: Any, installed: 3.1.2]
    │   └── MarkupSafe [required: >=2.0, installed: 2.1.2]
    ├── networkx [required: Any, installed: 3.0]
    ├── sympy [required: Any, installed: 1.12]
    │   └── mpmath [required: >=0.19, installed: 1.3.0]
    └── typing-extensions [required: Any, installed: 4.4.0]